arm_dpm: Fix resume when core state has been modified
Sometimes it is necessary to resume into a different state (ARM/Thumb)
than at debug state entry. According to the documentation this should
be possible with "arm core_state arm|thumb" before the resume command,
however the original code also restores the original CPSR, which
overrides whatever state the core was set to. The fix is to use the "BX"
instruction instead of "MOV pc, r0", which, when executed in debug
state, sets PC and core state according to the jump address. On exit
of the debug state, the CPU resumes at PC in the correct state.

Change-Id: Ic03153b4b250fbb8cf6c75f8e329fb34829aa35f
Signed-off-by: Matthias Welwarsky <matthias.welwarsky@sysgo.com>