board: add configuration for freescale twr-vf65gs10 system module
This file contains configuration for the twr-vf65gs10 system module
of the Freescale (NXP) tower system. It does not deal with the
CMSIS-DAP available via through USB but with the Cortex+ETM connector
on J5. The configuration also contains DDR and clock init code
hooked into the reset-init event handler.

Change-Id: I68303e0038e137dcadc57525e662428769fb69f2
Signed-off-by: Matthias Welwarsky <matthias.welwarsky@sysgo.com>